// enums/room-activity.ts
export enum RoomActivityType {
    ROOM_CREATED = "room_created",
    ROOM_UPDATED = "room_updated",
    MEMBER_JOINED = "member_joined",
    MEMBER_LEFT = "member_left",
    MEMBER_KICKED = "member_kicked",
    MEMBER_ROLE_UPDATED = "member_role_updated",
    INVOICE_CREATED = "invoice_created",
    INVOICE_UPDATED = "invoice_updated",
    INVOICE_DELETED = "invoice_deleted",
    INVOICE_PAID = "invoice_paid",
}